gtk_fb: add note about xshm surfaces
After some discussion with chergert on how this could be quicker
on X, it seemed worth noting at least something to this effect.
The details of doing this seemed annoying, because cairo apparently
doesn't just give you an xshm if appropriate via
gdk_window_create_similar_image_surface()

If that was all that were needed, this commit would just change the
cairo_image_surface_create() over to that.  But chergert claimed it
doesn't do the right thing here, so a comment it is until someone
cares enough to to investigate further.