gtk_fb: move window_{new,destroy} to fb_{init,shutdown}
Pages get allocated before fb_acquire, and when creating similar
pages to the underlying window it creates a depenency on the
window being available before acquire.  So move the window
create/destroy to the fb init and shutdown, acquire/release now
operate on the image within.

This commit also switches to using:
gdk_window_create_similar_image_surface()

for allocating pages, which may be more performant on backends
like X through the use of shared memory.