rototiller: bump submodule for rototiller_quiesce()
Since glimmer tears down and brings back up the world at will,
it needs a way to synchronize with idling the threaded renderer.

This may go away if librototiller gets refactored a bit to where
the main rendering loop moves into librototiller, but as long as
that's driven externally there needs to be a way to coordinate
at this level.