gtk_fb: s/configure-event/size-allocate/
The existing code made assumptions about the top-level window's
configure-event dimensions reflecting the underlying GtkImage's
dimensions.

This happened to be true most of the time on X11 desktops, but
client-side decorations have broken this assumption.

Instead resize the fb on size-allocate, and get the size
allocation out of the GtkImage widget *after* event propagation
finishes.

Thanks to chergert@redhat.com for assistance with quickly sorting
this out and providing the patch, and elektron@halo.nu reporting
the issue after testing on a librem5/phosh.