gtk_fb: when creating page surfaces don't inherit scaling
When calling gdk_window_create_similar_image_surface() I was
supplying '0' for the scaling factor which is just a blind
"inherit scaling from the window" request.

This is fine when the window isn't being scaled; everything is
just 1:1.

But when the window is scaled, and our output has been sized
according to the widget allocation (unscaled), these scaling
factors can't be the same if we intend for the page contents to
fill the scaled output.

Specify a scale of 1 instead, as that's how the allocation size
coordinates should be viewed, like normalized units.  The window
scale is how many times those units the window coordinates are
in.

This should fix highdpi/GDK_SCALE=2 usage.