glimmer: suppress re-rebuild from focus-out-event
When rebuilding the settings, an unnecessary focus-out-event is
produced that we don't want to queue another rebuild from.

In order to suppress the re-queueing of another rebuild, which
produces a sort of perpetual endless rebuild situation, use a
global flag for suppressing the rebuild queuing from the
focus-out-event callback.  The rebuild code sets this flag during
the rebuild, and the callback will skip queueing a rebuild when
it's true.

There's probably a more elegant solution, but I didn't see a
convenient way to suppress the focus events temporarily via
gtk/glib for the duration of the rebuild.

This works for now, and fixes the CPU burning behavior observed
on activated GtkEntry settings boxes.