rototiller: bump submodule for instantiated setups
Prior to this in some modules you could manipulate their settings
while watching their output, and the changes would be realized
*immediately*.

Rototiller has firmed up the application of settings so they're
no longer module-global and instead apply to an allocated setup
instance that is then supplied to create_context().

As a result, in glimmer, one must always now click "Go!" to see
the effects of new settings.